TimeTools: move stats duration formatter to tools
